+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.2
+- Temporary workaround while for webview resources are broken in a remote VS Code session. Awaiting fix from VS Code teams
+- Caching of external URLs for linked templates, large performance boost.
+
+
 ## 0.3.1
 - Export as PNG
 - Theme support (not fully implemented, only one theme currently 'original')